Rename FollowingObjects component for consistency
Renamed FollowingObjects.tsx to following-objects.tsx and updated import in dashboard page to use the new filename. This improves consistency in file naming conventions.
